Avoid findMember in early implicit candidate filtering.
We're only interested in whether or not a candidate implicit
type has a member with a given name, which we can determine
cheaply by checking if any ancestor has a so-named decl.

We avoid both the as-seen-froms that FindMember uses to
differentiate overriding and overloading, and creation of
overloaded symbols.