Cache ClassSymbol.javaBinaryNameString
The backend uses this string as a key to the map of BTypes,
and as such needs to call `javaBinaryNameString` for each method
or field reference in the code. Even though we have previously
optimized the creation of this string by bypassing the Name
abstraction and by correctly sizing string builders, we can
still speed things up by caching the resulting String on its
ClassSymbol.