Avoid excessive file stats during classfile writing
The existing implementation pessimistically checks that all
parent directories of the about-to-be-written class file are
indeed directories.

This commit bypasses this logic for the common case of writing
to a regular directory on disk, and optimistically assumes that
the parent directory exists. If an exception is thrown during
writing, it attempts to create the parent directory.

This still avoids a compiler crash if a parent directory is
actually a file, which is tested by the existing test,
`run/t5717.scala`.