Use a single repository in the bootstrap job
Recently, I changed the bootstap script to publish the "locker" build to scala-release-temp
and only the bootstrapped "quick" build to scala-integration. This commit reverts back to the
previous mechanism where "locker" is published to the same repo (scala-integration) and later
overwritten.

The reason is that we want to use scala-release-temp for publishing integration builds of commits
that are not (yet) merged to scala/scala. Such builds are useful for preliminary testing /
benchmarking of PRs / development revisions. This means that we want to add scala-release-temp
as a resolver in the community build and the benchmark job. If we have same-named releases in
both repos (one "locker", one "quick", so they are not the same), we would not know which one
is being picked up.

If we want to avoid the overwriting in the future, we could work on a solution that sets a different
version number for "locker", but we have to be careful when building the modules; maybe setting the
scalaBinaryVersion would be enough.