Improve serializability of lambdas that call methods in module
Depending on how it is written, a reference to an member of an
enclosing module can be either be qualified by `This(moduleClass)`
or by `Ident(module)`. The former was causing a unnecessary capture
of the module class.

This commit substitutes eligible `this` references when lifting
the closure body into a method.