Avoid excessive IO in classfile parser
The class file parser, used to read the java-defined classes from
the classpath, includes logic to search the output path for a .java
source file that corresponds to the value of the SourceFile attribute.

I haven't been able to figure out the rationale for that fix, but
it involves a non-neglible overhead, so this commits disables it
in the batch compiler.