Actually retract clashing synthetic apply/unapply
The completer set the IS_ERROR flag and I assumed the typer
dropped a synthetic tree with a symbol with that flag, because
the tree was not shown in -Xprint output.

It turns out, as explained by lrytz, that the mechanism was
fragile because it relied on the order in which completers are
run. We now cover both the case that:

  - the completer was run (and the `IS_ERROR` flag was set)
  before `addSynthetics` in `typedStat` iterates over the scope
  (since the symbol is already unlinked, the tree is not added,
  irrespective of its flags). For this case, we also remove the
  symbol from the synthetics in its unit.
  - the completer is triggered during the iteration in `addSynthetics`,
  which needs the check for the `IS_ERROR` flag during the iteration.

Thankfully, the community build caught my mistake, and lrytz provided
a good analysis and review.

Fix scala/bug#10261