Replace pattern matching with closure
Note: `{ (a, b) => ... }` is not the same as `{ case (a, b) => ... }`.
When used in a fold, the second form will auto-box the two arguments and then
take them apart with a pattern match. I have seen that pattern in other
places in the strawman as well (e.g. RedBlackTree), we should eliminate
it there as well to not lose performance.