Add uncheckedVariance annotation.
This is a case where the variance is indeed problematic, but scalac did
not detect it. Writing

    final class DupIterator(xs: Array[Iterable[T]]) ...

is morally equivalent to writing

    def dupIterator(xs: Array[Iterable[T]]) ...

In both cases `T` appears in non-variant position, so it can't be covariant in
class TrieIterator.

I guess the uncheckedVariace annotation is justified because DupIterator is only used
internally, but then we should declare it private.