Type annotation for ListBuffer
Without the annotation we get:

    |found:    strawman.collection.mutable.ListBuffer.To[
    |  util.Right[Nothing, String] | util.Left[Int, Nothing]
    |](xs3)
    |required: strawman.collection.mutable.ListBuffer'[scala.util.Either[Int, String]]

Unfortunately, with union types, we get a different type inference in dotty. I don't know
how to fix this without giving up on union types.

I also wonder why we see the `To` type. It's not good that this one leaks out into an error message.