Make Ordering.by create a new Ordering for performance
The previous Ordering.by method was a wrapper for Ordering.fromLessThan
that did not cache the values actually being compared.  This meant the
function had to be reevaluated if the first item was not less than the
second.  This reevaluation added a non-trivial performance cost to even
simple extractor functions like _.id.