Get rid of constrained with Ev.
It breaks Scala implicit inference, maybe because of the covariance/contravariance issue.
Not sure what it is, but in any case we only have a single instance, so abstraction is
premature (my rule of thumb is you need 3 plausible instances to make it worthwhile to
abstract).

So in this commit we replace constrained by ordered throughout.