Include the parts of a compound/refinement type in implicit scope.
According to the spec, the parts of a type T1 with ... with Tn { ... } are
the union of the parts of T1, ..., Tn. Without this, given an implicit view
from (A) => B defined in A's companion:

    val t: B = new A {}

and

    val t: B = { object a extends A {}; a }

had different behavior (only the latter compiled), which seemed surprising.

Now under an -Xsource:2.13 flag at retronym's request.

Fixes scala/bug#10288.