Inline Typer.isMonoContext
Contrary to the comment, it wasn't capturing the the context var
at typer initialization, which would not have been correct anyway.

The inlining here is motivated by avoiding allocation of lambda in
every Typer. Even when we're using this value in eta expansion,
SilentResult.filter will actually be inlined to avoid the need for a
lambda at all.