Enable -howtorun:repl -i -e
The runner concatenates -i files and -e expression
to run the code as a script. The addition of -I
broke detection of -i there, which is now fixed.

In addition, allow explicit `-howtorun:repl` which
is now useful for `-e`, so that REPL will handle
`-i` and `-I` and then evalute the expression as
usual (and as the `-help` text implies).

Fixes: scala/bug#10284