Explicit return type on Option.toRight/toLeft
Adds an explicit return type to the `toRight` and `toLeft` methods on
`Option`.

Note that before `Either` extended `Product with Serializable`
the inferred return type was `Product with Serializable with Either[A, B]`. (Side note: this return type was annoying as it messed with type-class based syntax enrichment.) But now that `Either` extends `Product with Serializable`, the inferred return type is `Either[A, B]`, so we can explicitly annotate it without breaking compatibility.