-Xlint:unused -Ywarn-unused is intuitive
The `-Xlint:unused` setting is just a boolean, but it
drives an expanding option on `-Ywarn-unused`.
Normally, expanding options can't be unset, so add
API for that.

This makes `-Xlint:_,-unused` work as advertised.
Other combinations such as `-Xlint -Ywarn-unused:-imports`
are also supported. That's because explicit unset trumps
an expanding option.