Optimize CommandLineParser to avoid exponential cost
Echoing the performance bug in the bash scripts (scala/bug#9279),
the command line parser exhibited exponential performance.

When using the command line arguments used to compile the standard
library, which contains ~ 700 source files, I noticed about 1% of
compilation time was being spend in evaluating the regex `(\S+)(.*)`.
Inspection of the parser reveals that this was repeatedly being
called on the rest of the command line as the parser consumed
each argument.

This commit switches to using a regex that only matches the
current argument, and then makes sure to split the input
string using structural sharing of the backing char[] (by using
`String.substring`, rather than high level Scala collections
operations from `StringLike#span`).

After this change, I no longer observed this area of code in the
top profiler results.