Optimize overriding pair cursor by caching the low member type
In #5865, I re-added caching of MethodSymbol.memberType, which would
make caching at this level redundant, but I was only able to re-add
it with for the typer phase. Overriding pairs are computed in refchecks
and erasure (for bridge generation and double-definition checks).

This commit caches the `pre.memberType(lowSymbol)` which is a win
when we must iterate through candidate `highSymbol`-s.

Compilation time of the standary library improves by 3% after this
change. I don't expect that benefit to generalize to much user code
which don't have the same deep inheritance heirarchy as the collections.