Revive part of TraverseTest
The commit shows that the fully generic part of TraverseTest works.
The rest is more tricky, but i am not convinced we need to support that.
I mean, we can always define optionSequence specialized to individual types -
there's not a lot to be gained to specify it separately for ordered
collections.