Implement :kind handling for type lambdas
When a type lambda is applied with concrete argument, for eample:

    def x: ({type l[A] = Either[Int, A]})#l[Nothing] = ???
    x: scala.util.Either[Int,Nothing]

The return type becomes an applied type with `Int` and `Nothing` supplied, and I was not able to retrieve the type lambda.
To workaround this issue, I am supplying `WildcardType` to represent the hole, calculate the kind as if no type parameters are supplied, and adjust the kind afterwards.