Improve presentation compiler with Shapeless macros
`-Ymacro-expand` discard was introduced to leave prefixes and
arguments of implicit applications in the typechecked tree in the
presentation compiler to facilite completion, type-at-cursor, etc
within those trees.

It seems that this mode interferes with implicit searches that involve
Shapeless's `mkLazy` macro.

This commit simply turns off the macro expansion discarding if
the currnt macro application is part of an application of implicit
arguments. There is no trees corresponding to source code in that
position, so we the concerns about IDE functionality are moot.

I couldn't disentangle the bug report from circe and shapeless,
so I've manually tested that the compiler from this commit
makes the following test project compile.

https://github.com/retronym/t9716