Drive bitmap computation from field, not accessor
Since bitmaps are only needed if there is a field, and the field carries
the transient annotation, don't go around via the accessor (we'd miss the
annotation).

Before we had a fields phase, this was tricky to do, but now it's
pretty straightforward to operate on field symbols, as we synthesize
the relevant ones, along with their bitmaps, during the fields phase
instead of waiting until mixin. As an extra bonus, the "artifact" fields
such as outer pointers etc don't exist yet, so we don't have to
exclude them.