Fix generation of derived value classes wrapping Unit, Null, and Nothing.
All sorts o' specialness going on here. Unit morphs into a BoxedUnit when
it's in a field, but void when it's the return type of a method, which is
expected. This means, though, that the unboxing method of a Unit-wrapping
value class has the signature `()V`, not `()Lscala/runtime/BoxedUnit`, so
attempting to use its value in the equals method spits out some wonderful
invalid bytecode, instead. Similar sadness occurs for Nothing and Null as
well.

The "solution" is to not even bother to check for equality, as we've only
got at most one legitimate value of each of these types. Because the code
is shared with `case class`es, this also changes the bytecode we generate
for them. Obviously this is an "unrelated change" as far as the bugs this
is meant to fix go, but it's innocuous enough as far as I can tell.

I also slipped a constructor call into the generated `ClassCastException`
that gets thrown when we are asked to emit a cast for a primitive type in
`BCodeBodyBuilder`, so we generate valid bytecode if we ever wind in that
branch. Discussion on scala/scala#5938 implies that this branch shouldn't
ever be reached, so add a devWarning now that it doesn't cause an obvious
error.

Fixes scala/bug#9240
Fixes scala/bug#10361