Fix check for varargs calling convention in overrides of Java methods.
The comment explains it all, but to repeat: Scala-defined varargs
methods take a `Seq` parameter; Java-defined varargs methods take
an `Array`. `transformArgs` in `uncurry` needs to wrap the passed
args in the right data structure. Checking for `fun.isJava` works
quite nicely, except for when we've made a superaccessor for some
Java-defined method, the superaccessor takes an `Array`, but does
not have `isJava` set, and therefore we wrap the args in a `Seq`,
netting us a tasty `ClassCastException` at runtime.

The solution: check with `isJavaVarArgsMethod` instead, which can
capture this distinction correctly.

Alternate solution: change SuperAccessors to generate a scala-ish
varargs method signature. I didn't pick this solution because, as
of right now, `superaccessors` knows not of varargs, and it might
be better left that way.

Fixes scala/bug#10368