Teach the optimizer about type tests on primitive box types.
Currently, we generate a box, `.isInstanceOf` call, and then unbox,
even for something as simple as:

    (1: Any) match { case i: Int => i; case _ => 0 }

The optimizer currently doesn't know enough to handle that case,
and emits naive code which `new`s up a box for `i`, checks if it
is truly a `java.lang.Integer`, then takes the value back out of
the box to return.

The simple fix for this involves replacing the `INSTANCEOF` node
in question with either a `ICONST_1` or an `ICONST_0`, depending
on whether or not the box is legit.