Fix regression under -Xcheckinit: param accessors don't need init checking
In #5936, some checks were removed that resulted in checkinit machinery
being synthesized for param accessor fields (ie, constructor vals).

This commit dials this back to the old behaviour, avoiding unnecessary
fields and code, and an unwanted VerifyError.

I've added a basic test case to show that checkinit errors are still
raised when we want them to be, and that the verify error is gone.