Avoid calls to static initializers in runtime reflection
Tested by having exception throwing <clinit> methods.

The Java reflection APIs offer the choice to the caller
via a boolean parameter to `Class.forName`. We don't have
such a parameter in our API, so the safest default is to
avoid calling the static initializers. The JVM will take
care of initialializing the class before any reflective
call is made, so this is safe to do.

Co-Authored-By: roman0yurin <gooood85@yandex.ru>