Translate appropriate JVM accessor flags into annotations.
a.k.a., once more into the ClassfileParser fray.

In `JavaMirrors`, add these synthetic annotations along with
the real ones (and the `@throws` annotation which we already
synthesize in the same place). In `ClassfileParser`, inspect
the access flags and apply the appropriate annotations then.
Happily, we were already doing The Right Thing[tm] for these
classes if we loaded their symbols via `JavaParsers`, so for
today that file escapes unscathed.

Fixes scala/bug#10042.