Structural invocation: find method by name if by signature fails
In a structural invocation, if the receiver object doesn't have
a method with the expected paramter types, check if there's a
single method matching the name. If so, invoke this method.

This fixes parts of https://github.com/scala/bug/issues/10334,
namely the parts that regressed with the new lambda encoding.
But the underlying issue still remains.