Disable tag driven release of the distribution
The bootstrap script is called within PR validation
and nightly builds, in addition to being used in the
release process.

It used to try to to automatically determine which
of these contexts was active based on the whether
HEAD was suitable tagged.

However, if same commit was rebuilt later as a nightly,
new binaries would be created and overwrite the official
ones.

This commit removes this logic. `SCALA_VER_BASE=...` or `publishToSonatype=yes` will need to be explicitly
provided in then environment to trigger publication.