Fix configuration of SBT incremental compiler
We want to use the (old) rules for incremental compilation
after editing files with macro definitions, as the new
rules lead to a lot of recompiles.

We'd previously configured this, but the change had been
somehow broken by a more recent change to the incremental
compiler options.

This commit moves both of the configs to the same part of
our build, which seems to make it stick:

```
% sbt consoleProject
> (incOptions in LocalRootProject).eval.recompileOnMacroDef
false
```