Fix bridging at the nexus of lambdas, dep. types, local modules
When we move the lambda RHS into a method (which is the target
of the LambdaMetaFactory invokedynamic bootstrap), we have to
substitute references to the lambda parameters with references
to the parameters of the lambda impl. method.

Turns out out substitution utility failed to do anything with
module classes info. The failure to substitute manifested as
an type mismatch between the inherited apply method and the one
defined in the companion factory method, which meant the requisite
bridge was skipped.