Correct scoping in typing application to block
Restrict a special case in typechecking applications to
the desugaring of right assoc operators, as in general
it's not sound to rewrite `{stats; expr}(a)` to `{stats; expr(a)}`,
because it changes visibility of names. Not sure why we even
need this special case, but that's a thread to pull on another day.

In the test case, the argument was binding to the inner
definition of `x`, which shouldn't have been in scope.

TODO: can we drop this when we eta-expand more aggressively?