Add infer Product with Serializable linter flag
This adds a new lint warning, -Xlint:infer-pws which warns when an
inferred value or method type is Product with Serializable. This
is oftentimes a programmer error from mixing up what would otherwise
be incompatible types (which could be caught by the -Xlint:infer-any
flag otherwise) that have the common PwS supertype. As this inferred
type tends to be useless, this lint flag helps avoid errors related
to this inference.

In addition, this also extends the same val and def inference check to
-Xlint:infer-any to make it consistent with the new linter setting.
Variable and method definitions that are inferred to be Any or AnyVal
will also add a warning unless they have explicit types ascribed.