Don't assume final member class in another comp. unit will stay final
The optimization in #5099 that avoided needless capture of the enclosing
class, and hence improved serializability of local classes and functions,
went too far. It also caused constructor calls of member classes to
use `null` rather than the actual outer reference as the `$outer`
constructor argument. The enclosed test case exhibits this problem
by witnessing an null `Outer.this`.

This commit limits the strategy to use `null`-s to the outer refererences
of anonymous and local classes, which rules out cross compilation unit
effects (in absence of `-opt`.)