Type Alternative patterns as the lub of their branches.
This came up on gitter a few weeks back, and the spec is
rather ambiguous on whether this happens or not, tersely
claiming that "All alternative patterns are type checked
with the expected type of the pattern". As such, this is
a bit of a spaghetti PR (in the "throw it at the wall to
see if it sticks" sense) motivated by checking Dotty and
seeing that The Future Of Scala™ duly types alternatives
in this way.

This will live behind a version flag because the sharper
type could accidentally change a public API if there was
no type ascription provided.

Also here is a tweak to weakLub so that it also uses the
numeric lub for singleton types of numbers. Otherwise we
get strange things like

    scala> def ct(v: Any) = ConstantType(Constant(v))
    ct: (v: Any)$r.intp.global.UniqueConstantType

    scala> val (one, twoL) = (ct(1), ct(2L))
    one: $r.intp.global.UniqueConstantType = Int(1)
    twoL: $r.intp.global.UniqueConstantType = Long(2L)

    scala> weakLub(one :: twoL :: Nil)
    res0: $r.intp.global.Type = AnyVal

    scala> weakLub((one :: twoL :: Nil) map (_.widen))
    res1: $r.intp.global.Type = Long

which just cannot be right.

The spec is now updated accordingly.

Review by adriaanm, retronym