Enable collection of all statistics
The following commit changes the previous val into a def so that
collection of all statistics is recorded. Before this change, most of
the statistics were disabled even if `-Ystatistics` was enabled because
the value of the `_enabled` was false when the object `Statistics` was
initialized. In the case of `-Ystatistics` being enabled, `_enabled` was
changed by the settings manager but that change never affected
`canEnable`.