Optimize statistics check via switchpoint
References: https://github.com/scala/scala-dev/issues/149.

It adds a mechanism to remove the overhead of checking for statistics in
the common case (where statistics are disabled). It does so by reusing a
mechanism proposed by Rémi Forax and brought to my attention by
Jason (more here: https://github.com/scala/scala-dev/issues/149).

This commit adds the following changes:

1. It adds the infrastructure so that statistics are optimized. This
means that now we have two different flags, `-Ystatistics` and
`-Yhot-statistics`, that enable different counters and timers. The hot
statistics are the ones that instrument key parts of the compiler
infrastructure and heavily affect the runtime (around 10%). The ones for
`-Ystatistics` enable most of the statistics, and are the ones that give
most information about the compiler and yet have a slight effect on
runtime.

2. It prints the statistics overhead per run, and does it through the
reporter instead of `println`.

The implementation of this commit has gone through several iterations
until performance has been deemed to be minimal, both for the cases
where statistics are enabled and disabled.

For a concrete analysis, see the PR discussion:
https://github.com/scala/scala/pull/6034.

There could still be some work optimizing the overhead where statistics
are enabled, but this commit does a pretty good job at it so far.