Print timers of phases at the end
The following commit adds the ability to print the timers for every
phase at the end of the `compileSources` execution. This is useful
because you can have an idea of how the running times of the phases are
distributed. It also works for phases injected by the incremental
compiler and compiler plugins (which allows devs to estimate how much
time do their compiler plugins take out of the total compile time).

It also removes the previous infrastructure to print these timings under
the verbose flag, and now reuses the full statistics infrastructure for
doing so.