Issue an error when trying to use a method as a qualifier.
The check for `implicitsEnabled` here seems older even than
that method; it appears to originally have looked at a flag
`reportGeneralErrors`, which seems a reasonable flag if you
want to check whether to report an error. An erroneous tree
that we do not error on, however, can be use as a qualifier
which will not necessarily cause an error until past typer,
when something finally notices the mistake and blows up.

Fixes scala/bug#10474.