By default, only run DCE on methods with an ATHROW.
GenBCode started to run a local DCE optimization on all methods
to afford simplicity in code generation of `Nothing`-typed
expressions.

Benchmarks show this was costing 1-2% of compile times.

This commit selectively run DCE on methods with the problematic
THROW, even under `-opt:l:none`. It then changes `-opt:l:default`
to the empty set of optimizations.

An existing test for this code gen quirk is updated to show with
either explicitly enabled or selective targeted DCE, clean code is
generated.