Primitive array types in ArrayOps / ArrayBuilder / ImmutableArray
This is similar to the old implementation. Primitive array types are not
polymorphic on the JVM so we need to special-case them in order to be
able to efficiently wrap existing arrays as collections.

We also remove ArrayView in favor of WrappedArray and ImmutableArray.
With the new equality semantics (where Views are never equal unless they
are the same reference) using a View as a default wrapper type for Array
is problematic. ImmutableArray is basically the same as ArrayView, except
it is an immutable.Seq and WrappedArray is the same for mutable.Seq, so
we get the expected equality semantics and strict operations for both
of them.

Implicit conversions from Array to Seq always produce a WrappedArray.
This is the safe choice because there is no guarantee that the array
is effectively immutable. ImmutableArray is intended for use in
compiler-generated varargs bridges and manually by the user via an
explicit conversion call.

Unlike the 2.12 collections the strawman currently requires a mutable
Seq to be Growable and Shrinkable. This needs to be changed in order to
implement WrappedArray. The new definition of mutable.Seq allows for
fixed-size collections like WrappedArray. A mutable.Seq which is
growable and shrinkable is a mutable.Buffer.

We also introduce mutable.IndexedSeq as a new type. It does not define
any new operations of its own but merely combines mutable.Seq with
collection.IndexedSeq for orthogonality with immutable.IndexedSeq.